Reorganize the structure (#13)
* switch to monai bundle config

* move yaml load function for readability

* user project import not working

* Functional user project importing, but ugly af

* unnecessary import statement

* import user project without yaml imports

* reorganize files, introduce contrib

* Remove contrib

* Add contrib module

* Rename iSSUE_TEMPLATE.md to ISSUE_TEMPLATE.md

* Place contrib into lighter

* rename importing module to dynamic_imports
