+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "lighter/contrib"]
+	path = lighter/contrib
+	url = https://github.com/project-lighter/contrib
