+++ setup.cfg
@@ -57,7 +57,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    lighter = lighter.cli:interface
+    lighter = lighter.utils.cli:interface
 
 [yapf]
 based_on_style = google
